Refactor - submodules names
